+++ gersemi/command_invocation_dumpers/multiple_signature_command_invocation_dumper.py
@@ -2,7 +2,8 @@ from contextlib import contextmanager
 from itertools import filterfalse
 from typing import Dict, List, Optional, Union
 from lark import Tree
-from gersemi.ast_helpers import is_comment, is_unquoted_argument
+from gersemi.ast_helpers import is_comment, is_one_of_keywords
+from gersemi.keywords import KeywordMatcher
 from gersemi.types import Nodes
 from .argument_aware_command_invocation_dumper import (
     ArgumentAwareCommandInvocationDumper,
@@ -26,7 +27,9 @@ def create_signature_patch(signature, old_class):
 
 
 class MultipleSignatureCommandInvocationDumper(ArgumentAwareCommandInvocationDumper):
-    customized_signatures: Dict[Optional[str], Dict[str, Union[List, Sections]]] = {}
+    customized_signatures: Dict[
+        Optional[KeywordMatcher], Dict[str, Union[List, Sections]]
+    ] = {}
 
     @contextmanager
     def _update_signature_characteristics(self, signature):
@@ -41,27 +44,35 @@ class MultipleSignatureCommandInvocationDumper(ArgumentAwareCommandInvocationDum
         finally:
             self.__class__ = old_class
 
+    def _get_signature_matcher(self, keyword):
+        for item in self.customized_signatures:
+            matcher = is_one_of_keywords([item])
+            if matcher(keyword):
+                return item
+        return None
+
     def format_command(self, tree):
         _, arguments = tree.children
         arguments = self._preprocess_arguments(arguments)
         arguments_only = filterfalse(is_comment, arguments.children)
         first_argument = next(arguments_only, None)
-        if first_argument is None or not is_unquoted_argument(first_argument):
+        matcher = self._get_signature_matcher(first_argument)
+
+        if matcher is None:
             signature = self.customized_signatures.get(None, None)
             if signature is None:
                 return super().format_command(tree)
             with self._update_signature_characteristics(signature):
                 return super().format_command(tree)
 
-        first_argument_as_value = first_argument.children[0]
-        signature = self.customized_signatures.get(first_argument_as_value, None)
+        signature = self.customized_signatures.get(matcher, None)
         if signature is None:
             signature = self.customized_signatures.get(None, None)
 
         with self._update_signature_characteristics(signature):
             return super().format_command(tree)
 
-    def _split_arguments(self, arguments: Nodes) -> List[Nodes]:
+    def _split_arguments(self, arguments: Nodes) -> Nodes:
         if len(arguments) == 0:
             return []
 
@@ -75,4 +86,8 @@ class MultipleSignatureCommandInvocationDumper(ArgumentAwareCommandInvocationDum
                     or (signature_node_as_value in self.multi_value_keywords)
                 ):
                     return super()._split_arguments(arguments)
-        return [[signature_node], *super()._split_arguments(rest)]
+
+        return [
+            Tree("positional_arguments", [signature_node]),
+            *super()._split_arguments(rest),
+        ]