+++ tests/formatter/dev.in.cmake
@@ -1,0 +1,1 @@
+project(Development scratch)